Enforce optional TCP connection limit (#1942)
As per https://github.com/SmingHub/Sming/pull/1942#issuecomment-553451304, this PR includes a connection count check before accepting a TCP connection. If a connection limit has been specified and accepting the new connection would exceed this value, then the connection is aborted.

Discussion #1940

+++ Sming/Core/Network/HttpServer.h
@@ -25,9 +25,9 @@
 #include "Http/HttpBodyParser.h"
 
 typedef struct {
-	int maxActiveConnections = 10; ///< maximum number of concurrent requests..
-	int keepAliveSeconds = 0;	  ///< default seconds to keep the connection alive before closing it
-	int minHeapSize = -1;		   ///< min heap size that is required to accept connection, -1 means use server default
+	uint16_t maxActiveConnections = 10; ///< maximum number of concurrent requests..
+	uint16_t keepAliveSeconds = 0;		///< default seconds to keep the connection alive before closing it
+	int minHeapSize = -1; ///< min heap size that is required to accept connection, -1 means use server default
 	bool useDefaultBodyParsers = 1; ///< if the default body parsers,  as form-url-encoded, should be used
 	bool closeOnContentError =
 		true; ///< close the connection if a body parser or resource fails to parse the body content.
